dpni: Reset # of multicast filters to default
Some DPNIs in the DPL only had one entry in the MAC filtering
table. This may become a problem with IPv6 support and in other cases
where multiple multicast addresses are assigned to the interface by
default.
Resetting the max_multicast_filters option to <0>, to allow
the MC to control this setting. By default it should be 64.

Note: Only modified DPNIs belonging to the Linux Ethernet driver, not
the NADK's.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>