Add DPCON objects to the RDB DPLs
Due to changes in the Ethernet driver, each Linux network interface
now requires a channel (DPCON) for each core/DPIO it runs on.
Update the RDB DPLs to reflect this change.

Also contains a few other small changes to the DPNI nodes properties:
* flow steering support: add DPNI_OPT_DIST_FS to the DPNI options
* increase the maximum key size to accomodate all possible header
fields which can be used for hashing/steering
* update distribution width: unintuitively, the value provided
here is incremented by one and rounded up to the next power of two,
so a value of 8 actually meant traffic is spread between 16 queues

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>