QDS: Fix DPL for SerDes 0x2a_0x49
Update the DPL for LS2085a QDS boards, for the 0x2a_0x49 SerDes and the
latest of Linux Ethernet driver.

The dependency on the SerDes comes from the 'connections' node, which
pairs specific DPNIs with hardcoded DPMACs and thereon (via the DPC
file for the same SerDes) with the external PHYs.

Changes include:
  - Added DPCON objects to support the current implementation of the
    Ethernet driver.
  - Re-computed the number of DPMCP objects as the number of DPNIs plus
    the number of DPMACs plus a fixed slack (10 DPMCPs) due to
    mc-bus (2), DPIO (1), restool (1), VFIO (2) and perhaps other future
    consumers.
  - Removed properties with default values (<0>) from the DPNI nodes.
  - Added some comments to sections in the DPL.

Note: this is a DPL example for a Linux-only container. As such, it
showcases the maximized number of DPNIs with maximal hash distribution
width. This consumes quite a lot of resources (DPCONs, most notably),
but various configurations can be adapted for less.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>