ls2080ardb: Convert more 8-WQ channels to 2-WQ channels
Convert 24 8-WQ channels into 96 2-WQ channels. The initial value of 1
was dimensioned strictly for the static DPL that only contains one DPNI,
but it would be good if the user were able to dynamically create many
more DPNIs without worrying about the number of DPCONs in the system.

This change effectively partitions the pool into
40 8-WQ DPCONs and 96 2-WQ DPCONs.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>