ls2085ardb, ls2080ardb: Match dpni max senders with rx hashing size
If Rx hashing is not enabled, Rx distribution size for DPAA2 Ethernet is
implicitly 1. The number of DPCONs in the DPL reflects this, but on the
other hand the max_senders attribute was too large and could not be
met by the available DPCONs.

Trim the max_senders value to match the reduced resourcing of the DPL.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>